refactor(mcp): own server lifecycle state machines
